+++ moroz/svc_preflight.go
@@ -1,0 +1,89 @@
+package moroz
+
+import (
+	"compress/zlib"
+	"context"
+	"encoding/json"
+	"net/http"
+	"time"
+
+	"github.com/go-kit/kit/endpoint"
+	"github.com/groob/moroz/santa"
+)
+
+func (svc *SantaService) Preflight(ctx context.Context, machineID string, p santa.PreflightPayload) (*santa.Preflight, error) {
+	config := svc.config(machineID)
+	pre := &santa.Preflight{
+		ClientMode:     toClientMode(config.ClientMode),
+		BlacklistRegex: config.BlacklistRegex,
+		WhitelistRegex: config.WhitelistRegex,
+		BatchSize:      config.BatchSize,
+	}
+	return pre, nil
+}
+
+func toClientMode(from string) santa.ClientMode {
+	switch from {
+	case "MONITOR":
+		return santa.Monitor
+	case "LOCKDOWN":
+		return santa.Lockdown
+	default:
+		return santa.Monitor
+	}
+}
+
+type preflightRequest struct {
+	MachineID string
+	payload   santa.PreflightPayload
+}
+
+type preflightResponse struct {
+	*santa.Preflight
+	Err error `json:"error,omitempty"`
+}
+
+func (r preflightResponse) Failed() error { return r.Err }
+
+func makePreflightEndpoint(svc santa.Service) endpoint.Endpoint {
+	return func(ctx context.Context, request interface{}) (interface{}, error) {
+		req := request.(preflightRequest)
+		preflight, err := svc.Preflight(ctx, req.MachineID, req.payload)
+		if err != nil {
+			return preflightResponse{Err: err}, nil
+		}
+		return preflightResponse{Preflight: preflight}, nil
+	}
+}
+
+func decodePreflightRequest(ctx context.Context, r *http.Request) (interface{}, error) {
+	zr, err := zlib.NewReader(r.Body)
+	if err != nil {
+		return nil, err
+	}
+	defer zr.Close()
+	defer r.Body.Close()
+	id, err := machineIDFromRequest(r)
+	if err != nil {
+		return nil, err
+	}
+	req := preflightRequest{MachineID: id}
+	if err := json.NewDecoder(zr).Decode(&req.payload); err != nil {
+		return nil, err
+	}
+	return req, nil
+}
+
+func (mw logmw) Preflight(ctx context.Context, machineID string, p santa.PreflightPayload) (pf *santa.Preflight, err error) {
+	defer func(begin time.Time) {
+		_ = mw.logger.Log(
+			"method", "Preflight",
+			"machine_id", machineID,
+			"err", err,
+			"took", time.Since(begin),
+		)
+	}(time.Now())
+
+	pf, err = mw.next.Preflight(ctx, machineID, p)
+	return
+}
