--- moroz/transport.go
@@ -1,30 +1,0 @@
-package moroz
-
-import (
-	"context"
-	"encoding/json"
-	"net/http"
-)
-
-type errorer interface {
-	error() error
-}
-
-func encodeResponse(ctx context.Context, w http.ResponseWriter, response interface{}) error {
-	if e, ok := response.(errorer); ok && e.error() != nil {
-		encodeError(ctx, e.error(), w)
-		return nil
-	}
-
-	enc := json.NewEncoder(w)
-	enc.SetIndent("", "  ")
-	return enc.Encode(response)
-}
-
-func encodeError(ctx context.Context, err error, w http.ResponseWriter) {
-	enc := json.NewEncoder(w)
-	enc.SetIndent("", "  ")
-	enc.Encode(map[string]string{
-		"error": err.Error(),
-	})
-}
